refactor(tasks): move Go1 shared base [skip ci]
