+++ src/unilab/envs/locomotion/__init__.py
@@ -1,7 +1,6 @@
 """Locomotion env registry bootstrap contract."""
 
 __unilab_registry_modules__ = (
-    "unilab.envs.locomotion.go1",
     "unilab.envs.locomotion.go2w",
     "unilab.envs.locomotion.g1",
     "unilab.envs.locomotion.go2_arm",
