--- src/unilab/envs/locomotion/go1/__init__.py
@@ -1,1 +1,0 @@
-"""Legacy Go1 shared base pending migration to :mod:`unilab.tasks`."""
